Extract RunResultGroup class to separate file
