+++ sarif/src/main/java/com/jetbrains/qodana/sarif/baseline/BaselineCalculation.java
@@ -1,11 +1,9 @@
 package com.jetbrains.qodana.sarif.baseline;
 
 import com.jetbrains.qodana.sarif.model.*;
-import kotlin.Pair;
 
 import java.util.*;
 import java.util.function.Function;
-import java.util.stream.StreamSupport;
 
 import static com.jetbrains.qodana.sarif.model.Result.BaselineState.*;
 
@@ -15,10 +13,10 @@ import static com.jetbrains.qodana.sarif.model.Result.BaselineState.*;
 public class BaselineCalculation {
     public static final String EQUAL_INDICATOR = "equalIndicator";
 
-    private int newResults = 0;
-    private int absentResults = 0;
-    private int unchangedResults = 0;
-    private final Options options;
+    int newResults = 0;
+    int absentResults = 0;
+    int unchangedResults = 0;
+    final Options options;
 
     private BaselineCalculation(Options options) {
         this.options = options;
@@ -59,7 +57,7 @@ public class BaselineCalculation {
 
             if (first.isPresent()) {
                 Run baselineRun = first.get();
-                new RunResultGroup(run, baselineRun).build();
+                new RunResultGroup(this, run, baselineRun).build();
                 baselineRuns.remove(baselineRun);
             } else {
                 unmatched.add(run);
@@ -73,13 +71,13 @@ public class BaselineCalculation {
                 markRunAsNew(run);
                 continue;
             }
-            new RunResultGroup(run, baselineRun).build();
+            new RunResultGroup(this, run, baselineRun).build();
         }
     }
 
     private void markRunAsNew(Run run) {
         for (Result result : run.getResults()) {
-            setBaselineState(result, NEW);
+            result.setBaselineState(NEW);
             newResults++;
         }
     }
@@ -95,9 +93,9 @@ public class BaselineCalculation {
     public static class Options {
         public static final Options DEFAULT = new Options();
 
-        private final boolean includeAbsent;
-        private final boolean includeUnchanged;
-        private final boolean fillBaselineState;
+        final boolean includeAbsent;
+        final boolean includeUnchanged;
+        final boolean fillBaselineState;
 
         /**
          * Provides information about incremental build.
@@ -111,7 +109,7 @@ public class BaselineCalculation {
          * Typically, wasChecked is true if result is in the scope of current check.
          */
         private static final Function<Result, Boolean> ALL_CHECKED = (result) -> true;
-        private final Function<Result, Boolean> wasChecked;
+        final Function<Result, Boolean> wasChecked;
 
         public Options() {
             includeAbsent = false;
@@ -152,110 +150,5 @@ public class BaselineCalculation {
         public boolean isFillBaselineState() {
             return fillBaselineState;
         }
-    }
-
-    private class RunResultGroup {
-        private final MultiMap<String, Result> baselineHashes;
-        private final MultiMap<String, Result> reportHashes;
-        private final MultiMap<ResultKey, Result> diffBaseline;
-        private final MultiMap<ResultKey, Result> diffReport;
-        private final Run report;
-        private final DescriptorLookup reportLookup;
-        private final DescriptorLookup baselineLookup;
-
-        public RunResultGroup(Run report, Run baseline) {
-            this.report = report;
-            this.reportLookup = new DescriptorLookup(report);
-            this.baselineLookup = new DescriptorLookup(baseline);
-
-            removeProblemsWithState(report, ABSENT);
-            Pair<MultiMap<String, Result>, MultiMap<ResultKey, Result>> reportIndices =
-                    BaselineKt.createIndices(report, m -> Collections.singletonList(m.getLastValue(EQUAL_INDICATOR)));
-
-            Pair<MultiMap<String, Result>, MultiMap<ResultKey, Result>> baselineIndices =
-                    BaselineKt.createIndices(baseline, m -> Collections.singletonList(m.getLastValue(EQUAL_INDICATOR)));
-            this.reportHashes = reportIndices.getFirst();
-            this.diffReport = reportIndices.getSecond();
-            this.baselineHashes = baselineIndices.getFirst();
-            this.diffBaseline = baselineIndices.getSecond();
-        }
-
-        private void removeProblemsWithState(Run report, Result.BaselineState state) {
-            report.getResults().removeIf(result -> result.getBaselineState() == state);
-        }
-
-        public void build() {
-            reportHashes.forEach(e -> {
-                if (baselineHashes.containsKey(e.getKey())) {
-                    e.getValue().forEach((it) -> setBaselineState(it, UNCHANGED));
-                    unchangedResults += e.getValue().size();
-                } else {
-                    e.getValue().forEach((it) -> diffReport.add(new ResultKey(it), it));
-                }
-            });
-
-            baselineHashes.forEach(e -> {
-                if (!reportHashes.containsKey(e.getKey())) {
-                    for (Result result : e.getValue()) {
-                        if (options.wasChecked.apply(result)) {
-                            diffBaseline.add(new ResultKey(result), result);
-                        } else {
-                            result.setBaselineState(UNCHANGED);
-                            report.getResults().add(result);
-                            unchangedResults += 1;
-                        }
-                    }
-                }
-            });
-
-            diffReport.forEach(e -> {
-                List<Result> baselineDiffBucket = diffBaseline.getOrEmpty(e.getKey());
-                for (Result result : e.getValue()) {
-                    if (baselineDiffBucket.isEmpty()) {
-                        setBaselineState(result, NEW);
-                        newResults++;
-                    } else {
-                        setBaselineState(result, UNCHANGED);
-                        baselineDiffBucket.remove(baselineDiffBucket.size() - 1);
-                        unchangedResults++;
-                    }
-                }
-            });
-
-            StreamSupport.stream(diffBaseline.spliterator(), false)
-                    .flatMap((it) -> it.getValue().stream())
-                    .forEach(result -> {
-                        if (options.wasChecked.apply(result)) {
-                            if (options.includeAbsent) {
-                                setBaselineState(result, ABSENT);
-                                absentResults++;
-                                report.getResults().add(result);
-                                if (reportLookup.findById(result.getRuleId()) == null) {
-                                    DescriptorWithLocation descriptor = baselineLookup.findById(result.getRuleId());
-                                    if (descriptor != null) descriptor.addTo(report);
-                                }
-                            }
-                        } else {
-                            result.setBaselineState(UNCHANGED);
-                            report.getResults().add(result);
-                            unchangedResults += 1;
-                        }
-                    });
-
-            if (!options.includeUnchanged) {
-                removeProblemsWithState(report, UNCHANGED);
-                unchangedResults = 0;
-            }
-
-            if (!options.fillBaselineState) {
-                for (Result result : report.getResults()) {
-                    result.setBaselineState(null);
-                }
-            }
-        }
-    }
-
-    private void setBaselineState(Result result, Result.BaselineState state) {
-        result.setBaselineState(state);
     }
 }
